Handle length with units + viewBox
1pt, 4em, 10mm, 100%, ... are all valid 'length' elements that are now
supported.
So far, we fixed DPI to 96 pixels per inch; 1em = 10px and 1ex = 5px to
begin with.
'%' unit should be accurate to the image dimension is specified;
otherwise screen is considered to be 800x600 pixels

viewBox is handled as a transformation of the newly introduced top_group
which groups any other elements in the structure.